appengine-with-cron/requirements.txt
google-api-python-client
GoogleAppEngineCloudStorageClient
SQLAlchemy
kubernetes
google-auth-httplib2
requests==2.3